Create endpoint list contacts
