+++ src/repositories/contact-repository.js
@@ -5,7 +5,18 @@ const createContact = async (contact) => {
     return storedContact
 }
 
+const listContact = async (page) => {
+    const limit = 10;
+    const offset = page * limit;
+    const { count, rows } = await Contact.findAndCountAll({
+        attributes: ['id', 'name', 'phone', 'email', 'message'],
+        offset: offset,
+        limit: limit,
+    })
+    return { total_contacts: count, contacts: rows }
+}
 
 module.exports = {
+    listContact,
     createContact
 }
